Move Housekeeping Common from stash to Github and fix packages
