+++ housekeeping-api/src/main/java/com/hotels/housekeeping/api/model/ILegacyReplicaPath.java
@@ -1,0 +1,35 @@
+/**
+ * Copyright (C) 2015-2017 Expedia Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.hotels.housekeeping.api.model;
+
+import java.io.Serializable;
+
+public interface ILegacyReplicaPath extends Serializable {
+
+  public String getEventId();
+
+  public String getPathEventId();
+
+  public String getPath();
+
+  public void setPath(String path);
+
+  public void setPathEventId(String pathEventId);
+
+  public void setEventId(String eventId);
+
+
+}
