+++ housekeeping-spring/src/main/resources/schema.sql
