--- src/test/resources/com/hotels/housekeeping/repository/legacyReplicaPath-after-save-test.xml
@@ -1,18 +1,0 @@
-<dataset>
-  <circus_train.legacy_replica_path
-              event_id="eventId_1"
-              path_event_id="eventId_0"
-              path="file:/foo/bar/0"/>
-  <circus_train.legacy_replica_path
-              event_id="eventId_2"
-              path_event_id="eventId_1"
-              path="file:/foo/bar/1"/>
-  <circus_train.legacy_replica_path
-              event_id="eventId_3"
-              path_event_id="eventId_2"
-              path="file:/foo/bar/2"/>
-  <circus_train.legacy_replica_path
-              event_id="eventId_4"
-              path_event_id="eventId_3"
-              path="file:/foo/bar/3"/>
-</dataset>
