--- src/test/resources/log4j.xml
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!--
-  Copyright (C) 2015-2017 Expedia Inc.
-
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
--->
-<!DOCTYPE log4j:configuration SYSTEM "log4j.dtd">
-<log4j:configuration xmlns:log4j='http://jakarta.apache.org/log4j/'>
-
-  <appender name="CONSOLE" class="org.apache.log4j.ConsoleAppender">
-    <layout class="org.apache.log4j.PatternLayout">
-      <param name="ConversionPattern" value="%d{ISO8601} %-5p %c:%L - %m%n" />
-    </layout>
-  </appender>
-
-  <logger name="com.hotels.bdp.circustrain">
-    <level value="DEBUG" />
-  </logger>
-  <logger name="org.apache.hadoop.tools">
-    <level value="DEBUG" />
-  </logger>
-
-  <root>
-    <level value="INFO" />
-    <appender-ref ref="CONSOLE" />
-  </root>
-
-</log4j:configuration>
